+++ gitutils/gitlab_utils.py
@@ -253,25 +253,35 @@ def create_group(group_name, description):
         raise gitutils_exception.GitutilsError(ex)
     return exitCode
 
+def print_search_output(group_indication, file_name, results):
+    if results:
+        for i in results:
+            print(const.bcolors.OKGREEN, file_name, const.bcolors.ENDC, ":")
+            print("\t\t Group: "+group_indication)
+            print("\t\t Project: "+i.get('project_name')+" (id "+ str(i.get('project_id'))+")")
+            print("\t\t Branch: "+i.get('branch'))
+            print("\t\t Weblink: "+i.get('webpath'))
+            print("\n")
+    else:
+        print(const.SEARCHFILE_EMPTY % (const.bcolors.FAIL, file_name, const.bcolors.ENDC))
+
 
 def find_file(file_name,group_indication):
     results =[]
     projects = get_group_projects(group_indication)
     for i in projects:
-        # print(i)
         # For every project's branch
         for b in i['branches']:
             # gets the project tree for the branch
             project_tree = get_project_tree(i.get('id'), b.name)
             for j in project_tree:
-                # print(j)
                 if file_name == j.get('name'):
-                    if j.get('type') == 'blob':
-                        results.append({
-                            'webpath':const.ENDPOINT+"/"+group_indication+"/"+i.get('name')+"/blob/"+b.name+"/"+j.get('path'),
-                            'branch':b.name,
-                            'project_id': i.get('id')
-                        })
+                    results.append({
+                        'webpath':const.ENDPOINT+"/"+group_indication+"/"+i.get('name')+"/"+j.get('type')+"/"+b.name+"/"+j.get('path'),
+                        'branch':b.name,
+                        'project_name':i.get('name'),
+                        'project_id': i.get('id')
+                    })
     return results
 
 def get_project_web_url(project_name):